Move do-system init logic from dat.py to do_fn.py
- Add create_do_manager() factory function in do_fn.py
- Add DoProxy class in do_fn.py for 'from dvc_dat import do'
- Simplify DatManager.__init__ to just call factory
- Remove backward-compat shims from DatManager.load/create
- Clean up __init__.py to just import from do_fn
